feat: remove cnofig.py
