+++ Makefile
@@ -3,7 +3,6 @@ test:
 	pytest src/kubernetes_cleanup_namespaces
 	pytest src/kubernetes_cleanup_pods
 	pytest src/kubernetes_configmap_jwt_scan
-	pytest src/kubernetes_export
 	pytest src/lib
 	pytest src/rabbitmq_export
 	pytest src/s3_prune_backups
