--- src/kubernetes_export/test/test_export.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-  pass
